init commit

+++ controllers/videoController.js
@@ -1,5 +1,6 @@
 import routes from "../routes"
 import Video from "../models/Video"
+import Comment from "../models/Comment"
 
 export const home = async (req, res) => {
   try {
@@ -29,6 +30,7 @@ export const getVideoUpload = (req, res) =>
   res.render("videoUpload", { pageTitle: "Upload video" })
 
 export const postVideoUpload = async (req, res) => {
+  console.log(req.body)
   const {
     body: { title, description },
     file: { path },
@@ -37,7 +39,10 @@ export const postVideoUpload = async (req, res) => {
     fileUrl: path,
     title,
     description,
+    creator: req.user.id,
   })
+  req.user.videos.push(newVideo.id)
+  req.user.save()
   res.redirect(routes.videoDetail(newVideo.id))
 }
 
@@ -47,6 +52,8 @@ export const videoDetail = async (req, res) => {
   } = req
   try {
     const video = await Video.findById(id)
+      .populate("creator")
+      .populate("comments")
     res.render("videoDetail", { pageTitle: video.title, video })
   } catch (e) {
     res.redirect(routes.home)
@@ -58,8 +65,13 @@ export const getEditVideo = async (req, res) => {
   } = req
   try {
     const video = await Video.findById(id)
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video })
+    if (video.creator != req.user.id) {
+      res.redirect(routes.home)
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video })
+    }
   } catch (e) {
+    console.log(e)
     res.redirect(routes.home)
   }
 }
@@ -81,9 +93,51 @@ export const deleteVideo = async (req, res) => {
     params: { id },
   } = req
   try {
-    await Video.findOneAndRemove({ _id: id })
+    const video = await Video.findById(id)
+    if (video.creator != req.user.id) {
+      res.redirect(routes.home)
+    } else {
+      await Video.findOneAndRemove({ _id: id })
+    }
   } catch (e) {
     console.log(e)
   }
   res.redirect(routes.home)
+}
+
+export const registerView = async (req, res) => {
+  const {
+    params: { id },
+  } = req
+  try {
+    const video = await Video.findById(id)
+    video.views += 1
+    video.save()
+    res.status(200)
+  } catch (e) {
+    res.status(400)
+  } finally {
+    res.end()
+  }
+}
+
+export const postAddComment = async (req, res) => {
+  const {
+    params: { id },
+    body: { comment },
+    user,
+  } = req
+  try {
+    const video = await Video.findById(id)
+    const newComment = await Comment.create({
+      text: comment,
+      creator: user.id,
+    })
+    video.comments.push(newComment.id)
+    video.save()
+  } catch (e) {
+    res.status(400)
+  } finally {
+    res.end()
+  }
 }
